Refactor App.js into smaller modules and components
- Created src/components, src/utils, and src/data directories.
- Extracted chart drawing, pattern analysis, and API fetching logic into src/utils.
- Moved static chart pattern data to src/data.
- Created several new React components in src/components:
  - PatternVisualization: Displays example chart patterns.
  - FlagIcon: Renders country flags.
  - StockSearch: Handles stock symbol input, suggestions, and fetching initiation.
  - ChartPreview: Displays the chart and the 'Analyze' button.
  - AnalysisResults: (Defined) To display analysis output.
- Updated App.js to use these new modules and components, significantly reducing its size and complexity.
- State and core handlers remain in App.js, passed as props.
