+++ src/App.js
@@ -1,556 +1,14 @@
 import React, { useState, useRef, useEffect } from 'react';
 import { AlertTriangle, TrendingUp, TrendingDown, Calendar, BarChart, Target, DollarSign, Search, RefreshCw, Clock } from 'lucide-react';
 import stocksData from './stocks.json';
-
-// Pattern drawing utility functions
-const drawLine = (ctx, points) => {
-  if (points.length < 2) return;
-  
-  ctx.beginPath();
-  ctx.moveTo(points[0][0], points[0][1]);
-  for (let i = 1; i < points.length; i++) {
-    ctx.lineTo(points[i][0], points[i][1]);
-  }
-  ctx.stroke();
-};
-
-const drawHeadAndShoulders = (ctx, margin, w, h) => {
-  const points = [
-    [margin, margin + h * 0.7],
-    [margin + w * 0.2, margin + h * 0.4],  // Left shoulder
-    [margin + w * 0.35, margin + h * 0.6], // Valley
-    [margin + w * 0.5, margin + h * 0.1],  // Head
-    [margin + w * 0.65, margin + h * 0.6], // Valley
-    [margin + w * 0.8, margin + h * 0.4],  // Right shoulder
-    [margin + w, margin + h * 0.8]
-  ];
-  drawLine(ctx, points);
-  
-  // Draw neckline
-  ctx.strokeStyle = '#dc2626';
-  ctx.lineWidth = 2;
-  ctx.setLineDash([5, 5]);
-  ctx.beginPath();
-  ctx.moveTo(margin + w * 0.35, margin + h * 0.6);
-  ctx.lineTo(margin + w * 0.65, margin + h * 0.6);
-  ctx.stroke();
-  ctx.setLineDash([]);
-};
-
-const drawInverseHeadAndShoulders = (ctx, margin, w, h) => {
-  const points = [
-    [margin, margin + h * 0.3],
-    [margin + w * 0.2, margin + h * 0.6],  // Left shoulder
-    [margin + w * 0.35, margin + h * 0.4], // Peak
-    [margin + w * 0.5, margin + h * 0.9],  // Head
-    [margin + w * 0.65, margin + h * 0.4], // Peak
-    [margin + w * 0.8, margin + h * 0.6],  // Right shoulder
-    [margin + w, margin + h * 0.2]
-  ];
-  drawLine(ctx, points);
-  
-  // Draw neckline
-  ctx.strokeStyle = '#10b981';
-  ctx.lineWidth = 2;
-  ctx.setLineDash([5, 5]);
-  ctx.beginPath();
-  ctx.moveTo(margin + w * 0.35, margin + h * 0.4);
-  ctx.lineTo(margin + w * 0.65, margin + h * 0.4);
-  ctx.stroke();
-  ctx.setLineDash([]);
-};
-
-const drawDoubleTop = (ctx, margin, w, h) => {
-  const points = [
-    [margin, margin + h * 0.7],
-    [margin + w * 0.25, margin + h * 0.2], // First peak
-    [margin + w * 0.4, margin + h * 0.6],  // Valley
-    [margin + w * 0.6, margin + h * 0.2],  // Second peak
-    [margin + w, margin + h * 0.8]
-  ];
-  drawLine(ctx, points);
-  
-  // Draw support line
-  ctx.strokeStyle = '#dc2626';
-  ctx.lineWidth = 2;
-  ctx.setLineDash([5, 5]);
-  ctx.beginPath();
-  ctx.moveTo(margin + w * 0.4, margin + h * 0.6);
-  ctx.lineTo(margin + w * 0.8, margin + h * 0.6);
-  ctx.stroke();
-  ctx.setLineDash([]);
-};
-
-const drawDoubleBottom = (ctx, margin, w, h) => {
-  const points = [
-    [margin, margin + h * 0.3],
-    [margin + w * 0.25, margin + h * 0.8], // First trough
-    [margin + w * 0.4, margin + h * 0.4],  // Peak
-    [margin + w * 0.6, margin + h * 0.8],  // Second trough
-    [margin + w, margin + h * 0.2]
-  ];
-  drawLine(ctx, points);
-  
-  // Draw resistance line
-  ctx.strokeStyle = '#10b981';
-  ctx.lineWidth = 2;
-  ctx.setLineDash([5, 5]);
-  ctx.beginPath();
-  ctx.moveTo(margin + w * 0.4, margin + h * 0.4);
-  ctx.lineTo(margin + w * 0.8, margin + h * 0.4);
-  ctx.stroke();
-  ctx.setLineDash([]);
-};
-
-const drawCupAndHandle = (ctx, margin, w, h) => {
-  // Cup
-  ctx.beginPath();
-  ctx.moveTo(margin, margin + h * 0.3);
-  ctx.quadraticCurveTo(margin + w * 0.35, margin + h * 0.8, margin + w * 0.7, margin + h * 0.3);
-  ctx.stroke();
-  
-  // Handle
-  const points = [
-    [margin + w * 0.7, margin + h * 0.3],
-    [margin + w * 0.8, margin + h * 0.45],
-    [margin + w * 0.9, margin + h * 0.4],
-    [margin + w, margin + h * 0.2]
-  ];
-  drawLine(ctx, points);
-};
-
-const drawAscendingTriangle = (ctx, margin, w, h) => {
-  const points = [
-    [margin, margin + h * 0.8],
-    [margin + w * 0.3, margin + h * 0.6],
-    [margin + w * 0.5, margin + h * 0.4],
-    [margin + w * 0.7, margin + h * 0.5],
-    [margin + w * 0.85, margin + h * 0.35],
-    [margin + w, margin + h * 0.2]
-  ];
-  drawLine(ctx, points);
-  
-  // Resistance line (horizontal)
-  ctx.strokeStyle = '#dc2626';
-  ctx.lineWidth = 2;
-  ctx.setLineDash([3, 3]);
-  ctx.beginPath();
-  ctx.moveTo(margin + w * 0.3, margin + h * 0.3);
-  ctx.lineTo(margin + w, margin + h * 0.3);
-  ctx.stroke();
-  
-  // Support line (ascending)
-  ctx.strokeStyle = '#10b981';
-  ctx.beginPath();
-  ctx.moveTo(margin, margin + h * 0.8);
-  ctx.lineTo(margin + w * 0.85, margin + h * 0.35);
-  ctx.stroke();
-  ctx.setLineDash([]);
-};
-
-const drawDescendingTriangle = (ctx, margin, w, h) => {
-  const points = [
-    [margin, margin + h * 0.2],
-    [margin + w * 0.3, margin + h * 0.4],
-    [margin + w * 0.5, margin + h * 0.6],
-    [margin + w * 0.7, margin + h * 0.5],
-    [margin + w * 0.85, margin + h * 0.65],
-    [margin + w, margin + h * 0.8]
-  ];
-  drawLine(ctx, points);
-  
-  // Support line (horizontal)
-  ctx.strokeStyle = '#10b981';
-  ctx.lineWidth = 2;
-  ctx.setLineDash([3, 3]);
-  ctx.beginPath();
-  ctx.moveTo(margin + w * 0.3, margin + h * 0.7);
-  ctx.lineTo(margin + w, margin + h * 0.7);
-  ctx.stroke();
-  
-  // Resistance line (descending)
-  ctx.strokeStyle = '#dc2626';
-  ctx.beginPath();
-  ctx.moveTo(margin, margin + h * 0.2);
-  ctx.lineTo(margin + w * 0.85, margin + h * 0.65);
-  ctx.stroke();
-  ctx.setLineDash([]);
-};
-
-const drawRisingWedge = (ctx, margin, w, h) => {
-  // Lower trend line (rising)
-  ctx.strokeStyle = '#10b981';
-  ctx.lineWidth = 2;
-  ctx.setLineDash([3, 3]);
-  ctx.beginPath();
-  ctx.moveTo(margin, margin + h * 0.8);
-  ctx.lineTo(margin + w, margin + h * 0.4);
-  ctx.stroke();
-  
-  // Upper trend line (rising slower)
-  ctx.strokeStyle = '#dc2626';
-  ctx.beginPath();
-  ctx.moveTo(margin, margin + h * 0.3);
-  ctx.lineTo(margin + w, margin + h * 0.2);
-  ctx.stroke();
-  ctx.setLineDash([]);
-  
-  // Price line
-  ctx.strokeStyle = '#2563eb';
-  ctx.lineWidth = 3;
-  const points = [
-    [margin, margin + h * 0.6],
-    [margin + w * 0.2, margin + h * 0.7],
-    [margin + w * 0.4, margin + h * 0.5],
-    [margin + w * 0.6, margin + h * 0.6],
-    [margin + w * 0.8, margin + h * 0.4],
-    [margin + w, margin + h * 0.3]
-  ];
-  drawLine(ctx, points);
-};
-
-const drawFallingWedge = (ctx, margin, w, h) => {
-  // Upper trend line (falling)
-  ctx.strokeStyle = '#dc2626';
-  ctx.lineWidth = 2;
-  ctx.setLineDash([3, 3]);
-  ctx.beginPath();
-  ctx.moveTo(margin, margin + h * 0.2);
-  ctx.lineTo(margin + w, margin + h * 0.6);
-  ctx.stroke();
-  
-  // Lower trend line (falling faster)
-  ctx.strokeStyle = '#10b981';
-  ctx.beginPath();
-  ctx.moveTo(margin, margin + h * 0.7);
-  ctx.lineTo(margin + w, margin + h * 0.8);
-  ctx.stroke();
-  ctx.setLineDash([]);
-  
-  // Price line
-  ctx.strokeStyle = '#2563eb';
-  ctx.lineWidth = 3;
-  const points = [
-    [margin, margin + h * 0.4],
-    [margin + w * 0.2, margin + h * 0.3],
-    [margin + w * 0.4, margin + h * 0.5],
-    [margin + w * 0.6, margin + h * 0.4],
-    [margin + w * 0.8, margin + h * 0.6],
-    [margin + w, margin + h * 0.7]
-  ];
-  drawLine(ctx, points);
-};
-
-const drawFlag = (ctx, margin, w, h) => {
-  // Strong move up (flagpole)
-  const points1 = [
-    [margin, margin + h * 0.9],
-    [margin + w * 0.4, margin + h * 0.2]
-  ];
-  drawLine(ctx, points1);
-  
-  // Consolidation (flag)
-  const points2 = [
-    [margin + w * 0.4, margin + h * 0.2],
-    [margin + w * 0.5, margin + h * 0.3],
-    [margin + w * 0.6, margin + h * 0.25],
-    [margin + w * 0.7, margin + h * 0.35],
-    [margin + w * 0.8, margin + h * 0.3],
-    [margin + w, margin + h * 0.1]
-  ];
-  drawLine(ctx, points2);
-  
-  // Flag boundaries
-  ctx.strokeStyle = '#dc2626';
-  ctx.lineWidth = 1;
-  ctx.setLineDash([2, 2]);
-  ctx.beginPath();
-  ctx.moveTo(margin + w * 0.4, margin + h * 0.2);
-  ctx.lineTo(margin + w * 0.8, margin + h * 0.25);
-  ctx.stroke();
-  ctx.beginPath();
-  ctx.moveTo(margin + w * 0.4, margin + h * 0.35);
-  ctx.lineTo(margin + w * 0.8, margin + h * 0.4);
-  ctx.stroke();
-  ctx.setLineDash([]);
-};
-
-const drawGenericPattern = (ctx, margin, w, h) => {
-  const points = [
-    [margin, margin + h * 0.5],
-    [margin + w * 0.2, margin + h * 0.3],
-    [margin + w * 0.4, margin + h * 0.7],
-    [margin + w * 0.6, margin + h * 0.4],
-    [margin + w * 0.8, margin + h * 0.6],
-    [margin + w, margin + h * 0.2]
-  ];
-  drawLine(ctx, points);
-};
-
-const drawPattern = (ctx, pattern, w, h) => {
-  const margin = 20;
-  const chartW = w - 2 * margin;
-  const chartH = h - 2 * margin;
-  
-  // Set default styles
-  ctx.strokeStyle = '#2563eb';
-  ctx.lineWidth = 3;
-  ctx.lineCap = 'round';
-  ctx.lineJoin = 'round';
-  
-  // Draw grid
-  ctx.strokeStyle = '#f0f0f0';
-  ctx.lineWidth = 1;
-  for (let i = 0; i <= 4; i++) {
-    const y = margin + (i / 4) * chartH;
-    ctx.beginPath();
-    ctx.moveTo(margin, y);
-    ctx.lineTo(w - margin, y);
-    ctx.stroke();
-  }
-  
-  // Reset line style for pattern
-  ctx.strokeStyle = '#2563eb';
-  ctx.lineWidth = 3;
-  
-  switch (pattern) {
-    case 'head-and-shoulders':
-      drawHeadAndShoulders(ctx, margin, chartW, chartH);
-      break;
-    case 'inverse-head-and-shoulders':
-      drawInverseHeadAndShoulders(ctx, margin, chartW, chartH);
-      break;
-    case 'double-top':
-      drawDoubleTop(ctx, margin, chartW, chartH);
-      break;
-    case 'double-bottom':
-      drawDoubleBottom(ctx, margin, chartW, chartH);
-      break;
-    case 'cup-and-handle':
-      drawCupAndHandle(ctx, margin, chartW, chartH);
-      break;
-    case 'ascending-triangle':
-      drawAscendingTriangle(ctx, margin, chartW, chartH);
-      break;
-    case 'descending-triangle':
-      drawDescendingTriangle(ctx, margin, chartW, chartH);
-      break;
-    case 'wedge-rising':
-      drawRisingWedge(ctx, margin, chartW, chartH);
-      break;
-    case 'wedge-falling':
-      drawFallingWedge(ctx, margin, chartW, chartH);
-      break;
-    case 'flag':
-      drawFlag(ctx, margin, chartW, chartH);
-      break;
-    default:
-      drawGenericPattern(ctx, margin, chartW, chartH);
-  }
-  
-  // Add pattern name
-  ctx.fillStyle = '#1f2937';
-  ctx.font = 'bold 12px Inter, Arial, sans-serif';
-  ctx.textAlign = 'center';
-  ctx.fillText(pattern.split('-').map(word => 
-    word.charAt(0).toUpperCase() + word.slice(1)
-  ).join(' '), w / 2, h - 5);
-};
-
-// Pattern Visualization Component
-const PatternVisualization = ({ patternName, width = 300, height = 150 }) => {
-  const canvasRef = useRef(null);
-
-  useEffect(() => {
-    if (!canvasRef.current || !patternName) return;
-    
-    const canvas = canvasRef.current;
-    const ctx = canvas.getContext('2d');
-    
-    // Set canvas size
-    canvas.width = width;
-    canvas.height = height;
-    
-    // Clear canvas
-    ctx.fillStyle = '#ffffff';
-    ctx.fillRect(0, 0, width, height);
-    
-    // Draw pattern based on type
-    drawPattern(ctx, patternName, width, height);
-  }, [patternName, width, height]);
-
-  return (
-    <canvas 
-      ref={canvasRef}
-      style={{ 
-        border: '1px solid #e5e7eb', 
-        borderRadius: '8px',
-        background: '#ffffff',
-        maxWidth: '100%',
-        height: 'auto'
-      }}
-    />
-  );
-};
-
-// Cross-platform flag component (most reliable solution)
-const FlagIcon = ({ country, size = 16 }) => {
-  const height = Math.round(size * 0.75); // 4:3 aspect ratio
-  
-  if (country === 'India') {
-    return (
-      <svg width={size} height={height} viewBox="0 0 16 12" style={{ marginRight: '4px', verticalAlign: 'middle' }}>
-        <rect width="16" height="4" fill="#ff9933"/>
-        <rect y="4" width="16" height="4" fill="white"/>
-        <rect y="8" width="16" height="4" fill="#138808"/>
-        <circle cx="8" cy="6" r="1.5" stroke="#000080" strokeWidth="0.3" fill="none"/>
-        <g stroke="#000080" strokeWidth="0.2">
-          <line x1="6.5" y1="6" x2="9.5" y2="6"/>
-          <line x1="8" y1="4.5" x2="8" y2="7.5"/>
-          <line x1="6.8" y1="5.2" x2="9.2" y2="6.8"/>
-          <line x1="9.2" y1="5.2" x2="6.8" y2="6.8"/>
-        </g>
-      </svg>
-    );
-  } else {
-    return (
-      <svg width={size} height={height} viewBox="0 0 16 12" style={{ marginRight: '4px', verticalAlign: 'middle' }}>
-        <rect width="16" height="12" fill="#bf0a30"/>
-        <rect y="1" width="16" height="1" fill="white"/>
-        <rect y="3" width="16" height="1" fill="white"/>
-        <rect y="5" width="16" height="1" fill="white"/>
-        <rect y="7" width="16" height="1" fill="white"/>
-        <rect y="9" width="16" height="1" fill="white"/>
-        <rect y="11" width="16" height="1" fill="white"/>
-        <rect width="6" height="6" fill="#002868"/>
-      </svg>
-    );
-  }
-};
-
-// Enhanced chart patterns with detailed analysis
-const chartPatterns = {
-  'head-and-shoulders': {
-    description: 'A bearish reversal pattern with three peaks, the middle being the highest',
-    prediction: 'down',
-    timeframe: '7-21 days',
-    daysDown: '10-25 days',
-    daysUp: '0 days',
-    reliability: 85,
-    recommendation: 'sell',
-    entryStrategy: 'Sell on neckline break below',
-    exitStrategy: 'Target: Height of head below neckline',
-    breakoutDays: '3-7 days'
-  },
-  'inverse-head-and-shoulders': {
-    description: 'A bullish reversal pattern with three troughs, the middle being the lowest',
-    prediction: 'up',
-    timeframe: '7-21 days',
-    daysDown: '0 days',
-    daysUp: '14-30 days',
-    reliability: 83,
-    recommendation: 'buy',
-    entryStrategy: 'Buy on neckline break above',
-    exitStrategy: 'Target: Height of head above neckline',
-    breakoutDays: '2-6 days'
-  },
-  'double-top': {
-    description: 'A bearish reversal pattern showing two distinct peaks at similar price levels',
-    prediction: 'down',
-    timeframe: '14-28 days',
-    daysDown: '14-35 days',
-    daysUp: '0 days',
-    reliability: 78,
-    recommendation: 'sell',
-    entryStrategy: 'Sell on break below valley',
-    exitStrategy: 'Target: Distance between peaks and valley',
-    breakoutDays: '5-10 days'
-  },
-  'double-bottom': {
-    description: 'A bullish reversal pattern showing two distinct troughs at similar price levels',
-    prediction: 'up',
-    timeframe: '14-28 days',
-    daysDown: '0 days',
-    daysUp: '21-42 days',
-    reliability: 79,
-    recommendation: 'buy',
-    entryStrategy: 'Buy on break above peak',
-    exitStrategy: 'Target: Distance between troughs and peak',
-    breakoutDays: '4-8 days'
-  },
-  'cup-and-handle': {
-    description: 'A bullish continuation pattern resembling a cup followed by a short downward trend',
-    prediction: 'up',
-    timeframe: '30-60 days',
-    daysDown: '0 days',
-    daysUp: '30-90 days',
-    reliability: 88,
-    recommendation: 'buy',
-    entryStrategy: 'Buy on handle breakout',
-    exitStrategy: 'Target: Cup depth above breakout',
-    breakoutDays: '7-14 days'
-  },
-  'ascending-triangle': {
-    description: 'A bullish continuation pattern with a flat upper resistance and rising lower support',
-    prediction: 'up',
-    timeframe: '21-35 days',
-    daysDown: '0 days',
-    daysUp: '14-45 days',
-    reliability: 72,
-    recommendation: 'buy',
-    entryStrategy: 'Buy on resistance breakout',
-    exitStrategy: 'Target: Triangle height above breakout',
-    breakoutDays: '3-9 days'
-  },
-  'descending-triangle': {
-    description: 'A bearish continuation pattern with a flat lower support and falling upper resistance',
-    prediction: 'down',
-    timeframe: '21-35 days',
-    daysDown: '21-42 days',
-    daysUp: '0 days',
-    reliability: 74,
-    recommendation: 'sell',
-    entryStrategy: 'Sell on support breakdown',
-    exitStrategy: 'Target: Triangle height below breakdown',
-    breakoutDays: '4-11 days'
-  },
-  'flag': {
-    description: 'A short-term consolidation pattern that typically continues the prior trend',
-    prediction: 'continuation',
-    timeframe: '7-14 days',
-    daysDown: '5-10 days',
-    daysUp: '5-14 days',
-    reliability: 68,
-    recommendation: 'hold',
-    entryStrategy: 'Buy/Sell on flag breakout',
-    exitStrategy: 'Target: Flagpole height in breakout direction',
-    breakoutDays: '1-5 days'
-  },
-  'wedge-rising': {
-    description: 'A bearish reversal pattern with converging upward trending lines',
-    prediction: 'down',
-    timeframe: '14-28 days',
-    daysDown: '14-35 days',
-    daysUp: '0 days',
-    reliability: 76,
-    recommendation: 'sell',
-    entryStrategy: 'Sell on lower trendline break',
-    exitStrategy: 'Target: Wedge height below break',
-    breakoutDays: '6-12 days'
-  },
-  'wedge-falling': {
-    description: 'A bullish reversal pattern with converging downward trending lines',
-    prediction: 'up',
-    timeframe: '14-28 days',
-    daysDown: '0 days',
-    daysUp: '14-35 days',
-    reliability: 77,
-    recommendation: 'buy',
-    entryStrategy: 'Buy on upper trendline break',
-    exitStrategy: 'Target: Wedge height above break',
-    breakoutDays: '5-10 days'
-  }
-};
+// import { drawPattern } from './utils/chartDrawing'; // drawPattern is used by PatternVisualization, not directly here
+import PatternVisualization from './components/PatternVisualization';
+import FlagIcon from './components/FlagIcon';
+import { chartPatterns } from './data/chartPatterns';
+import { detectPatternFromPriceData, calculateBreakoutTiming } from './utils/patternAnalysis';
+import { fetchYahooFinanceData, generateMockStockData } from './utils/stockApi';
+import StockSearch from './components/StockSearch';
+import ChartPreview from './components/ChartPreview';
 
 function StockChartAnalyzer() {
   // Extract data from imported JSON
@@ -578,355 +36,6 @@ function StockChartAnalyzer() {
   const canvasRef = useRef(null);
   const chartCanvasRef = useRef(null);
   const inputRef = useRef(null);
-
-  // Enhanced pattern detection using actual price data
-  const detectPatternFromPriceData = (prices) => {
-    if (!prices || prices.length < 20) return null;
-
-    const closes = prices.map(p => p.close);
-    const highs = prices.map(p => p.high);
-    const lows = prices.map(p => p.low);
-    
-    // Find significant peaks and troughs
-    const peaks = findPeaksAndTroughs(highs, true);
-    const troughs = findPeaksAndTroughs(lows, false);
-    
-    // Calculate various technical indicators
-    const rsi = calculateRSI(closes, 14);
-    const currentRSI = rsi[rsi.length - 1];
-    
-    const sma20 = calculateSMA(closes, 20);
-    const sma50 = calculateSMA(closes, 50);
-    
-    const currentPrice = closes[closes.length - 1];
-    const priceVsSMA20 = ((currentPrice - sma20[sma20.length - 1]) / sma20[sma20.length - 1]) * 100;
-    const priceVsSMA50 = ((currentPrice - sma50[sma50.length - 1]) / sma50[sma50.length - 1]) * 100;
-    
-    // Enhanced pattern recognition
-    const patternData = analyzePatterns(peaks, troughs, closes, highs, lows);
-    
-    return {
-      pattern: patternData.pattern,
-      confidence: calculateDynamicConfidence(patternData, currentRSI, priceVsSMA20, priceVsSMA50),
-      technicals: {
-        rsi: currentRSI,
-        priceVsSMA20,
-        priceVsSMA50,
-        peaks: peaks.length,
-        troughs: troughs.length
-      }
-    };
-  };
-
-  // Find peaks and troughs in price data
-  const findPeaksAndTroughs = (data, isPeak = true) => {
-    const results = [];
-    const lookback = 5; // Look 5 periods in each direction
-    
-    for (let i = lookback; i < data.length - lookback; i++) {
-      let isSignificant = true;
-      
-      for (let j = i - lookback; j <= i + lookback; j++) {
-        if (j === i) continue;
-        
-        if (isPeak) {
-          if (data[j] >= data[i]) {
-            isSignificant = false;
-            break;
-          }
-        } else {
-          if (data[j] <= data[i]) {
-            isSignificant = false;
-            break;
-          }
-        }
-      }
-      
-      if (isSignificant) {
-        results.push({ index: i, value: data[i] });
-      }
-    }
-    
-    return results;
-  };
-
-  // Calculate RSI
-  const calculateRSI = (data, period = 14) => {
-    const gains = [];
-    const losses = [];
-    
-    for (let i = 1; i < data.length; i++) {
-      const change = data[i] - data[i - 1];
-      gains.push(change > 0 ? change : 0);
-      losses.push(change < 0 ? Math.abs(change) : 0);
-    }
-    
-    const rsi = [];
-    
-    for (let i = period - 1; i < gains.length; i++) {
-      const avgGain = gains.slice(i - period + 1, i + 1).reduce((a, b) => a + b) / period;
-      const avgLoss = losses.slice(i - period + 1, i + 1).reduce((a, b) => a + b) / period;
-      
-      if (avgLoss === 0) {
-        rsi.push(100);
-      } else {
-        const rs = avgGain / avgLoss;
-        rsi.push(100 - (100 / (1 + rs)));
-      }
-    }
-    
-    return rsi;
-  };
-
-  // Calculate Simple Moving Average
-  const calculateSMA = (data, period) => {
-    const sma = [];
-    
-    for (let i = period - 1; i < data.length; i++) {
-      const sum = data.slice(i - period + 1, i + 1).reduce((a, b) => a + b);
-      sma.push(sum / period);
-    }
-    
-    return sma;
-  };
-
-  // Analyze patterns with enhanced logic
-  const analyzePatterns = (peaks, troughs, closes, highs, lows) => {
-    const recentData = closes.slice(-30); // Last 30 days
-    const priceRange = Math.max(...recentData) - Math.min(...recentData);
-    const tolerance = priceRange * 0.03; // 3% tolerance
-    
-    // Head and Shoulders detection
-    if (peaks.length >= 3) {
-      const lastThreePeaks = peaks.slice(-3);
-      const [left, head, right] = lastThreePeaks;
-      
-      if (head.value > left.value + tolerance && head.value > right.value + tolerance) {
-        const necklineConfidence = Math.abs(left.value - right.value) / tolerance;
-        if (necklineConfidence <= 2) { // Peaks should be relatively equal
-          return { pattern: 'head-and-shoulders', strength: 0.8 + (1 - necklineConfidence / 2) * 0.2 };
-        }
-      }
-    }
-
-    // Inverse Head and Shoulders
-    if (troughs.length >= 3) {
-      const lastThreeTroughs = troughs.slice(-3);
-      const [left, head, right] = lastThreeTroughs;
-      
-      if (head.value < left.value - tolerance && head.value < right.value - tolerance) {
-        const necklineConfidence = Math.abs(left.value - right.value) / tolerance;
-        if (necklineConfidence <= 2) {
-          return { pattern: 'inverse-head-and-shoulders', strength: 0.75 + (1 - necklineConfidence / 2) * 0.25 };
-        }
-      }
-    }
-
-    // Double Top
-    if (peaks.length >= 2) {
-      const lastTwoPeaks = peaks.slice(-2);
-      const [first, second] = lastTwoPeaks;
-      
-      if (Math.abs(first.value - second.value) <= tolerance) {
-        return { pattern: 'double-top', strength: 0.7 + (1 - Math.abs(first.value - second.value) / tolerance) * 0.3 };
-      }
-    }
-
-    // Double Bottom
-    if (troughs.length >= 2) {
-      const lastTwoTroughs = troughs.slice(-2);
-      const [first, second] = lastTwoTroughs;
-      
-      if (Math.abs(first.value - second.value) <= tolerance) {
-        return { pattern: 'double-bottom', strength: 0.65 + (1 - Math.abs(first.value - second.value) / tolerance) * 0.35 };
-      }
-    }
-
-    // Triangle patterns
-    const trianglePattern = detectTrianglePatterns(peaks, troughs, closes);
-    if (trianglePattern) return trianglePattern;
-
-    // Cup and Handle (simplified)
-    if (detectCupAndHandle(closes)) {
-      return { pattern: 'cup-and-handle', strength: 0.8 };
-    }
-
-    // Flag pattern
-    const flagPattern = detectFlagPattern(closes);
-    if (flagPattern) return flagPattern;
-
-    // Wedge patterns
-    const wedgePattern = detectWedgePatterns(peaks, troughs, closes);
-    if (wedgePattern) return wedgePattern;
-
-    // Default to flag if no clear pattern
-    return { pattern: 'flag', strength: 0.4 };
-  };
-
-  // Detect triangle patterns
-  const detectTrianglePatterns = (peaks, troughs, closes) => {
-    if (peaks.length < 2 || troughs.length < 2) return null;
-
-    const recentPeaks = peaks.slice(-3);
-    const recentTroughs = troughs.slice(-3);
-
-    // Ascending triangle (horizontal resistance, rising support)
-    if (recentPeaks.length >= 2) {
-      const peakTrend = (recentPeaks[recentPeaks.length - 1].value - recentPeaks[0].value) / recentPeaks.length;
-      const troughTrend = recentTroughs.length >= 2 ? 
-        (recentTroughs[recentTroughs.length - 1].value - recentTroughs[0].value) / recentTroughs.length : 0;
-
-      if (Math.abs(peakTrend) < 0.5 && troughTrend > 0.5) {
-        return { pattern: 'ascending-triangle', strength: 0.7 };
-      }
-
-      // Descending triangle (falling resistance, horizontal support)
-      if (peakTrend < -0.5 && Math.abs(troughTrend) < 0.5) {
-        return { pattern: 'descending-triangle', strength: 0.7 };
-      }
-    }
-
-    return null;
-  };
-
-  // Detect cup and handle pattern
-  const detectCupAndHandle = (closes) => {
-    if (closes.length < 50) return false;
-
-    const recent50 = closes.slice(-50);
-    const firstThird = recent50.slice(0, 16);
-    const middleThird = recent50.slice(16, 33);
-    const lastThird = recent50.slice(33);
-
-    const firstAvg = firstThird.reduce((a, b) => a + b) / firstThird.length;
-    const middleAvg = middleThird.reduce((a, b) => a + b) / middleThird.length;
-    const lastAvg = lastThird.reduce((a, b) => a + b) / lastThird.length;
-
-    // Cup: decline then recovery
-    if (middleAvg < firstAvg * 0.9 && lastAvg > firstAvg * 0.95) {
-      return true;
-    }
-
-    return false;
-  };
-
-  // Detect flag pattern
-  const detectFlagPattern = (closes) => {
-    if (closes.length < 15) return null;
-
-    const recent15 = closes.slice(-15);
-    const range = Math.max(...recent15) - Math.min(...recent15);
-    const avgPrice = recent15.reduce((a, b) => a + b) / recent15.length;
-
-    // Flag: tight consolidation
-    if (range / avgPrice < 0.05) { // Less than 5% range
-      return { pattern: 'flag', strength: 0.6 };
-    }
-
-    return null;
-  };
-
-  // Detect wedge patterns
-  const detectWedgePatterns = (peaks, troughs, closes) => {
-    if (peaks.length < 2 || troughs.length < 2) return null;
-
-    const recentPeaks = peaks.slice(-3);
-    const recentTroughs = troughs.slice(-3);
-
-    if (recentPeaks.length >= 2 && recentTroughs.length >= 2) {
-      const peakTrend = (recentPeaks[recentPeaks.length - 1].value - recentPeaks[0].value);
-      const troughTrend = (recentTroughs[recentTroughs.length - 1].value - recentTroughs[0].value);
-
-      // Rising wedge (both lines rising, converging)
-      if (peakTrend > 0 && troughTrend > 0 && troughTrend > peakTrend * 0.5) {
-        return { pattern: 'wedge-rising', strength: 0.65 };
-      }
-
-      // Falling wedge (both lines falling, converging)
-      if (peakTrend < 0 && troughTrend < 0 && Math.abs(troughTrend) > Math.abs(peakTrend) * 0.5) {
-        return { pattern: 'wedge-falling', strength: 0.65 };
-      }
-    }
-
-    return null;
-  };
-
-  // Calculate dynamic confidence based on multiple factors
-  const calculateDynamicConfidence = (patternData, rsi, priceVsSMA20, priceVsSMA50) => {
-    let baseConfidence = chartPatterns[patternData.pattern]?.reliability || 70;
-    let patternStrength = patternData.strength || 0.5;
-    
-    // Adjust confidence based on pattern strength
-    let confidence = baseConfidence * patternStrength;
-    
-    // Technical indicator adjustments
-    if (patternData.pattern.includes('up') || patternData.pattern === 'ascending-triangle' || 
-        patternData.pattern === 'inverse-head-and-shoulders' || patternData.pattern === 'double-bottom') {
-      // Bullish patterns
-      if (rsi > 30 && rsi < 70) confidence += 5; // Good RSI range
-      if (priceVsSMA20 > 0) confidence += 3; // Above 20-day MA
-      if (priceVsSMA50 > 0) confidence += 3; // Above 50-day MA
-    } else if (patternData.pattern.includes('down') || patternData.pattern === 'descending-triangle' ||
-               patternData.pattern === 'head-and-shoulders' || patternData.pattern === 'double-top') {
-      // Bearish patterns
-      if (rsi > 30 && rsi < 70) confidence += 5;
-      if (priceVsSMA20 < 0) confidence += 3; // Below 20-day MA
-      if (priceVsSMA50 < 0) confidence += 3; // Below 50-day MA
-    }
-    
-    // Volume and data quality adjustments
-    confidence += Math.random() * 10 - 5; // Add some variation
-    
-    return Math.max(45, Math.min(92, Math.round(confidence)));
-  };
-
-  // Calculate breakout timing
-  const calculateBreakoutTiming = (patternName, stockData, confidence) => {
-    const pattern = chartPatterns[patternName];
-    if (!pattern || !stockData) return null;
-
-    const baseBreakoutDays = pattern.breakoutDays || '3-7 days';
-    const breakoutRange = baseBreakoutDays.split('-');
-    const minDays = parseInt(breakoutRange[0]);
-    const maxDays = parseInt(breakoutRange[1]);
-
-    // Adjust timing based on confidence and market conditions
-    let adjustedMin = minDays;
-    let adjustedMax = maxDays;
-
-    if (confidence > 80) {
-      // High confidence = faster breakout
-      adjustedMin = Math.max(1, minDays - 1);
-      adjustedMax = Math.max(adjustedMin + 1, maxDays - 2);
-    } else if (confidence < 60) {
-      // Low confidence = slower breakout
-      adjustedMin = minDays + 1;
-      adjustedMax = maxDays + 3;
-    }
-
-    // Calculate actual dates
-    const today = new Date();
-    const minBreakoutDate = new Date(today);
-    const maxBreakoutDate = new Date(today);
-    
-    minBreakoutDate.setDate(today.getDate() + adjustedMin);
-    maxBreakoutDate.setDate(today.getDate() + adjustedMax);
-
-    return {
-      daysRange: `${adjustedMin}-${adjustedMax} days`,
-      minDate: minBreakoutDate.toLocaleDateString('en-US', { 
-        weekday: 'short', 
-        month: 'short', 
-        day: 'numeric' 
-      }),
-      maxDate: maxBreakoutDate.toLocaleDateString('en-US', { 
-        weekday: 'short', 
-        month: 'short', 
-        day: 'numeric' 
-      }),
-      confidence: confidence > 75 ? 'High' : confidence > 60 ? 'Medium' : 'Low'
-    };
-  };
 
   // Type-ahead functionality with market support
   const filterSuggestions = (input) => {
@@ -1060,111 +169,6 @@ function StockChartAnalyzer() {
         <span key={index} style={{ backgroundColor: '#fef3c7', fontWeight: '600' }}>{part}</span> : 
         part
     );
-  };
-
-  const fetchYahooFinanceData = async (symbol) => {
-    try {
-      // Using a CORS proxy service to bypass CORS restrictions
-      const proxyUrl = 'https://api.allorigins.win/raw?url=';
-      const yahooUrl = encodeURIComponent(`https://query1.finance.yahoo.com/v8/finance/chart/${symbol}?range=3mo&interval=1d`);
-      
-      const response = await fetch(proxyUrl + yahooUrl);
-      
-      if (!response.ok) {
-        throw new Error(`HTTP error! status: ${response.status}`);
-      }
-      
-      const data = await response.json();
-      
-      if (data.chart?.error) {
-        throw new Error(data.chart.error.description || 'Invalid stock symbol');
-      }
-      
-      if (!data.chart?.result?.[0]) {
-        throw new Error('No data found for this symbol');
-      }
-      
-      const result = data.chart.result[0];
-      const timestamps = result.timestamp;
-      const quotes = result.indicators.quote[0];
-      const meta = result.meta;
-      
-      const prices = timestamps.map((timestamp, index) => ({
-        date: new Date(timestamp * 1000).toISOString().split('T')[0],
-        open: quotes.open[index],
-        high: quotes.high[index],
-        low: quotes.low[index],
-        close: quotes.close[index],
-        volume: quotes.volume[index]
-      })).filter(price => price.close !== null && price.close !== undefined);
-      
-      if (prices.length === 0) {
-        throw new Error('No valid price data found');
-      }
-      
-      return {
-        symbol: symbol.toUpperCase(),
-        companyName: meta.longName || symbol,
-        currency: meta.currency || 'USD',
-        exchange: meta.exchangeName || '',
-        currentPrice: meta.regularMarketPrice || prices[prices.length - 1].close,
-        prices: prices
-      };
-    } catch (error) {
-      console.error('Yahoo Finance API Error:', error);
-      
-      // Fallback: Generate realistic mock data if API fails
-      if (error.message.includes('CORS') || error.message.includes('fetch')) {
-        return generateMockStockData(symbol);
-      }
-      
-      throw new Error(`Failed to fetch data for ${symbol}: ${error.message}`);
-    }
-  };
-
-  // Generate realistic mock data as fallback
-  const generateMockStockData = (symbol) => {
-    const isIndianStock = symbol.includes('.NS');
-    const basePrice = isIndianStock ? Math.random() * 2000 + 500 : Math.random() * 200 + 50; // Higher base price for Indian stocks in INR
-    const prices = [];
-    let currentPrice = basePrice;
-    
-    // Generate 90 days of realistic price data
-    for (let i = 89; i >= 0; i--) {
-      const date = new Date();
-      date.setDate(date.getDate() - i);
-      
-      const volatility = 0.02;
-      const change = (Math.random() - 0.5) * 2 * volatility;
-      currentPrice = currentPrice * (1 + change);
-      
-      const open = currentPrice;
-      const close = currentPrice * (1 + (Math.random() - 0.5) * 0.01);
-      const high = Math.max(open, close) * (1 + Math.random() * 0.005);
-      const low = Math.min(open, close) * (1 - Math.random() * 0.005);
-      const volume = Math.floor(Math.random() * 10000000) + 1000000;
-      
-      prices.push({
-        date: date.toISOString().split('T')[0],
-        open: parseFloat(open.toFixed(2)),
-        high: parseFloat(high.toFixed(2)),
-        low: parseFloat(low.toFixed(2)),
-        close: parseFloat(close.toFixed(2)),
-        volume: volume
-      });
-      
-      currentPrice = close;
-    }
-    
-    return {
-      symbol: symbol.toUpperCase(),
-      companyName: isIndianStock ? `${symbol.replace('.NS', '')} Ltd.` : `${symbol.toUpperCase()} Inc.`,
-      currency: isIndianStock ? 'INR' : 'USD',
-      exchange: isIndianStock ? 'NSE' : 'NASDAQ',
-      currentPrice: currentPrice,
-      prices: prices,
-      isMockData: true
-    };
   };
 
   // Create enhanced chart image from stock data
@@ -1482,198 +486,25 @@ function StockChartAnalyzer() {
         </div>
       </div>
 
-      {/* Stock Symbol Input with Type-ahead */}
-      <div style={{ marginBottom: '32px' }}>
-        <label style={{ display: 'block', fontWeight: '600', marginBottom: '12px', color: '#1a202c', fontSize: '18px' }}>
-          <Search size={20} style={{ display: 'inline', marginRight: '8px', verticalAlign: 'middle' }} />
-          Get Live Stock Chart (3-Month Analysis)
-        </label>
-        
-        <div style={{ position: 'relative', marginBottom: '16px' }}>
-          <div style={{ display: 'flex', gap: '12px', flexWrap: 'wrap' }}>
-            <div style={{ flex: '1', minWidth: '300px', position: 'relative' }}>
-              <input
-                ref={inputRef}
-                type="text"
-                value={stockSymbol}
-                onChange={(e) => handleInputChange(e.target.value)}
-                onKeyDown={handleKeyDown}
-                onFocus={handleInputFocus}
-                onBlur={handleInputBlur}
-                placeholder="🔍 Search: AAPL, TCS.NS, Reliance, Microsoft, HDFC Bank..."
-                style={{ 
-                  width: '100%', 
-                  padding: '14px 16px', 
-                  border: showSuggestions ? '2px solid #6366f1' : '2px solid rgba(99, 102, 241, 0.2)', 
-                  borderRadius: showSuggestions ? '8px 8px 0 0' : '8px',
-                  fontSize: '16px', 
-                  fontWeight: '500', 
-                  outline: 'none', 
-                  transition: 'border-color 0.2s',
-                  borderBottom: showSuggestions ? '1px solid rgba(99, 102, 241, 0.2)' : '2px solid rgba(99, 102, 241, 0.2)'
-                }}
-              />
-              
-              {/* Suggestions Dropdown */}
-              {showSuggestions && (
-                <div style={{
-                  position: 'absolute',
-                  top: '100%',
-                  left: '0',
-                  right: '0',
-                  backgroundColor: 'white',
-                  border: '2px solid #6366f1',
-                  borderTop: 'none',
-                  borderRadius: '0 0 8px 8px',
-                  boxShadow: '0 4px 20px rgba(0, 0, 0, 0.15)',
-                  zIndex: 1000,
-                  maxHeight: '300px',
-                  overflowY: 'auto'
-                }}>
-                  {filteredSuggestions.length > 0 ? (
-                    filteredSuggestions.map((stock, index) => (
-                      <div
-                        key={stock.symbol}
-                        onClick={() => selectSuggestion(stock)}
-                        style={{
-                          padding: '12px 16px',
-                          cursor: 'pointer',
-                          backgroundColor: index === selectedSuggestionIndex ? '#f3f4f6' : 'white',
-                          borderBottom: index < filteredSuggestions.length - 1 ? '1px solid #e5e7eb' : 'none',
-                          transition: 'background-color 0.2s'
-                        }}
-                        onMouseEnter={() => setSelectedSuggestionIndex(index)}
-                      >
-                        <div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
-                          <div>
-                            <div style={{ fontWeight: '600', fontSize: '15px', color: '#1f2937' }}>
-                              {highlightMatch(stock.symbol, stockSymbol)}
-                            </div>
-                            <div style={{ fontSize: '13px', color: '#6b7280', marginTop: '2px' }}>
-                              {highlightMatch(stock.name, stockSymbol)}
-                            </div>
-                          </div>
-                          <div style={{ display: 'flex', gap: '4px', alignItems: 'center' }}>
-                            <div style={{ 
-                              fontSize: '10px', 
-                              color: stock.market === 'India' ? '#dc2626' : '#2563eb', 
-                              backgroundColor: stock.market === 'India' ? '#fef2f2' : '#eff6ff', 
-                              padding: '2px 6px', 
-                              borderRadius: '4px',
-                              fontWeight: '600',
-                              border: `1px solid ${stock.market === 'India' ? '#fecaca' : '#dbeafe'}`,
-                              display: 'flex',
-                              alignItems: 'center',
-                              gap: '2px'
-                            }}>
-                              <FlagIcon country={stock.market} size={12} />
-                              {stock.market === 'India' ? 'NSE' : 'US'}
-                            </div>
-                            <div style={{ 
-                              fontSize: '11px', 
-                              color: '#9ca3af', 
-                              backgroundColor: '#f3f4f6', 
-                              padding: '2px 6px', 
-                              borderRadius: '4px',
-                              fontWeight: '500'
-                            }}>
-                              {stock.sector}
-                            </div>
-                          </div>
-                        </div>
-                      </div>
-                    ))
-                  ) : stockSymbol.length >= 1 ? (
-                    <div style={{ 
-                      padding: '16px', 
-                      textAlign: 'center', 
-                      color: '#6b7280', 
-                      fontSize: '14px' 
-                    }}>
-                      <div style={{ marginBottom: '8px' }}>🔍 No stocks found</div>
-                      <div style={{ fontSize: '12px' }}>
-                        Try searching by symbol (AAPL) or company name (Apple)
-                      </div>
-                    </div>
-                  ) : null}
-                </div>
-              )}
-            </div>
-            
-            <button
-              onClick={() => stockSymbol.trim() && fetchStockData(stockSymbol.toUpperCase())}
-              disabled={loading || !stockSymbol.trim()}
-              style={{ 
-                padding: '14px 24px', 
-                background: loading ? '#9ca3af' : 'linear-gradient(135deg, #6366f1 0%, #8b5cf6 100%)', 
-                color: 'white', 
-                border: 'none', 
-                borderRadius: '8px', 
-                fontWeight: '600', 
-                cursor: loading ? 'not-allowed' : 'pointer', 
-                display: 'flex', 
-                alignItems: 'center', 
-                gap: '8px', 
-                transition: 'all 0.2s',
-                minWidth: '140px',
-                justifyContent: 'center'
-              }}
-            >
-              {loading ? <RefreshCw size={16} style={{ animation: 'spin 1s linear infinite' }} /> : <Search size={16} />}
-              {loading ? 'Fetching...' : 'Get Chart'}
-            </button>
-          </div>
-        </div>
-
-        {/* Popular stocks with market indicators */}
-        <div>
-          <p style={{ fontSize: '14px', color: '#4a5568', marginBottom: '12px', fontWeight: '500' }}>
-            Popular Stocks from {stockDatabase.length}+ available (<FlagIcon country="US" size={12} />US + <FlagIcon country="India" size={12} />Indian Markets):
-          </p>
-          <div style={{ display: 'flex', gap: '8px', flexWrap: 'wrap' }}>
-            {popularStocksData.map(stock => (
-              <button
-                key={stock.symbol}
-                onClick={() => selectStock(stock.symbol)}
-                disabled={loading}
-                style={{ 
-                  padding: '8px 12px', 
-                  background: stockSymbol === stock.symbol ? 'linear-gradient(135deg, #6366f1 0%, #8b5cf6 100%)' : 'rgba(99, 102, 241, 0.1)', 
-                  color: stockSymbol === stock.symbol ? 'white' : '#4f46e5', 
-                  border: '1px solid rgba(99, 102, 241, 0.3)', 
-                  borderRadius: '20px', 
-                  fontSize: '13px', 
-                  fontWeight: '500', 
-                  cursor: loading ? 'not-allowed' : 'pointer', 
-                  transition: 'all 0.2s',
-                  opacity: loading ? 0.6 : 1,
-                  display: 'flex',
-                  alignItems: 'center',
-                  gap: '4px'
-                }}
-                onMouseEnter={(e) => {
-                  if (stockSymbol !== stock.symbol && !loading) {
-                    e.target.style.background = 'rgba(99, 102, 241, 0.2)';
-                  }
-                }}
-                onMouseLeave={(e) => {
-                  if (stockSymbol !== stock.symbol && !loading) {
-                    e.target.style.background = 'rgba(99, 102, 241, 0.1)';
-                  }
-                }}
-              >
-                <FlagIcon country={stock.market} size={12} />
-                {stock.symbol.replace('.NS', '')}
-              </button>
-            ))}
-          </div>
-          
-          {/* Quick market examples */}
-          <div style={{ marginTop: '12px', fontSize: '12px', color: '#6b7280' }}>
-            <strong>Examples:</strong> Search from {stockDatabase.length}+ stocks - try "TCS" (Indian IT), "Reliance" (Indian Oil), "AAPL" (US Tech), "HDFC" (Indian Banking), "NVDA" (US Semiconductors), or "Wipro" (Indian IT)
-          </div>
-        </div>
-      </div>
+      <StockSearch
+        stockSymbol={stockSymbol}
+        handleInputChange={handleInputChange}
+        handleKeyDown={handleKeyDown}
+        handleInputFocus={handleInputFocus}
+        handleInputBlur={handleInputBlur}
+        showSuggestions={showSuggestions}
+        filteredSuggestions={filteredSuggestions}
+        selectedSuggestionIndex={selectedSuggestionIndex}
+        setSelectedSuggestionIndex={setSelectedSuggestionIndex}
+        selectSuggestion={selectSuggestion}
+        loading={loading}
+        fetchStockData={fetchStockData}
+        popularStocksData={popularStocksData}
+        selectStock={selectStock}
+        highlightMatch={highlightMatch}
+        inputRef={inputRef}
+        stockDatabase={stockDatabase}
+      />
 
       {/* Error Display */}
       {error && (
@@ -1710,41 +541,12 @@ function StockChartAnalyzer() {
         />
       </div>
       
-      {uploadedImage && (
-        <div style={{ marginBottom: '32px' }}>
-          <div style={{ width: '100%', height: '400px', background: 'rgba(255, 255, 255, 0.9)', borderRadius: '16px', overflow: 'hidden', marginBottom: '20px', display: 'flex', alignItems: 'center', justifyContent: 'center', border: '2px solid rgba(0, 0, 0, 0.1)', boxShadow: '0 4px 20px rgba(0, 0, 0, 0.1)' }}>
-            <img 
-              src={uploadedImage} 
-              alt="Stock chart" 
-              style={{ maxWidth: '100%', maxHeight: '100%', objectFit: 'contain', borderRadius: '12px' }}
-            />
-          </div>
-          
-          {stockData && (
-            <div style={{ background: 'linear-gradient(135deg, rgba(16, 185, 129, 0.1), rgba(52, 211, 153, 0.1))', border: '2px solid rgba(16, 185, 129, 0.3)', borderRadius: '12px', padding: '16px', marginBottom: '16px', fontSize: '15px', color: '#065f46' }}>
-              <div style={{ fontWeight: '700', marginBottom: '8px' }}>📊 Stock Information (3-Month Data):</div>
-              <div><strong>Symbol:</strong> {stockData.symbol} | <strong>Company:</strong> {stockData.companyName}</div>
-              <div><strong>Current Price:</strong> {stockData.currency === 'INR' || stockData.symbol.includes('.NS') ? '₹' : '$'}{stockData.currentPrice?.toFixed(2)} {stockData.currency} | <strong>Data Points:</strong> {stockData.prices.length} days</div>
-              {stockData.isMockData && <div style={{ color: '#f59e0b', fontStyle: 'italic', marginTop: '4px' }}>⚠️ Using demo data - API temporarily unavailable</div>}
-            </div>
-          )}
-          
-          <button
-            onClick={analyzeChart}
-            disabled={loading}
-            style={{ width: '100%', background: loading ? '#9ca3af' : 'linear-gradient(135deg, #6366f1 0%, #8b5cf6 100%)', color: 'white', border: 'none', padding: '18px 24px', fontSize: '18px', fontWeight: '600', borderRadius: '12px', cursor: loading ? 'not-allowed' : 'pointer', textTransform: 'uppercase', letterSpacing: '0.5px', transition: 'all 0.3s', boxShadow: loading ? 'none' : '0 4px 20px rgba(99, 102, 241, 0.4)' }}
-          >
-            {loading ? (
-              <span style={{ display: 'flex', alignItems: 'center', justifyContent: 'center', gap: '8px' }}>
-                <RefreshCw size={20} style={{ animation: 'spin 1s linear infinite' }} />
-                Analyzing Pattern...
-              </span>
-            ) : (
-              '🔍 Analyze Chart Pattern'
-            )}
-          </button>
-        </div>
-      )}
+      <ChartPreview
+        uploadedImage={uploadedImage}
+        stockData={stockData}
+        analyzeChart={analyzeChart}
+        loading={loading}
+      />
       
       {/* Results Section */}
       {prediction && patternDetected && (
